requirements.txt
aniso8601==9.0.1
argcomplete==2.0.0
asgiref==3.6.0
autopep8==1.6.0
certifi==2022.12.7
click==8.1.3
distlib==0.3.6
Django==4.1.7
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-graphql-jwt==0.3.4
django-js-asset==1.2.2
django-webpack-loader==1.8.1
django_debug_toolbar==3.8.1
filelock==3.9.0
graphene==3.2.1
graphene-django==3.0.0
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
packaging==23.0
Pillow==9.4.0
pipenv==2023.2.18
pipx==1.1.0
platformdirs==3.0.0
promise==2.3
psycopg2-binary==2.9.5
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2018.7
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
userpath==1.8.0
virtualenv==20.20.0
virtualenv-clone==0.5.7
